remove test package and unused methods (#755)
