+++ pkg/blockatlas/tx_test.go
@@ -1,14 +1,12 @@
-package main
+package blockatlas
 
 import (
 	"github.com/stretchr/testify/assert"
 	"github.com/trustwallet/blockatlas/coin"
-	"github.com/trustwallet/blockatlas/observer"
-	"github.com/trustwallet/blockatlas/pkg/blockatlas"
 	"testing"
 )
 
-var transferDst1 = blockatlas.Tx{
+var transferDst1 = Tx{
 	ID:     "1681EE543FB4B5A628EF21D746E031F018E226D127044A4F9BA5EE2542A44555",
 	Coin:   coin.BNB,
 	From:   "tbnb1fhr04azuhcj0dulm7ka40y0cqjlafwae9k9gk2",
@@ -16,33 +14,16 @@ var transferDst1 = blockatlas.Tx{
 	Fee:    "125000",
 	Date:   1555049867,
 	Block:  7761368,
-	Status: blockatlas.StatusCompleted,
+	Status: StatusCompleted,
 	Memo:   "test",
-	Meta: blockatlas.Transfer{
+	Meta: Transfer{
 		Value:    "10000000000000",
 		Decimals: 8,
 		Symbol:   "BNB",
 	},
 }
 
-var transferDst2 = blockatlas.Tx{
-	ID:     "1681EE543FB4B5A628EF21D746E031F018E226D127044A4F9BA5EE2542A44556",
-	Coin:   coin.BNB,
-	From:   "tbnb1fhr04azuhcj0dulm7ka40y0cqjlafwae9k9gk2",
-	To:     "tbnb1fhr04azuhcj0dulm7ka40y0cqjlafwae9k9gk2",
-	Fee:    "125000",
-	Date:   1555049867,
-	Block:  7761368,
-	Status: blockatlas.StatusCompleted,
-	Memo:   "test",
-	Meta: blockatlas.Transfer{
-		Value:    "10000000000000",
-		Decimals: 8,
-		Symbol:   "BNB",
-	},
-}
-
-var nativeTransferDst1 = blockatlas.Tx{
+var nativeTransferDst1 = Tx{
 	ID:     "95CF63FAA27579A9B6AF84EF8B2DFEAC29627479E9C98E7F5AE4535E213FA4C9",
 	Coin:   coin.BNB,
 	From:   "tbnb1ttyn4csghfgyxreu7lmdu3lcplhqhxtzced45a",
@@ -50,9 +31,9 @@ var nativeTransferDst1 = blockatlas.Tx{
 	Fee:    "125000",
 	Date:   1555117625,
 	Block:  7928667,
-	Status: blockatlas.StatusCompleted,
+	Status: StatusCompleted,
 	Memo:   "test",
-	Meta: blockatlas.NativeTokenTransfer{
+	Meta: NativeTokenTransfer{
 		TokenID:  "YLC-D8B",
 		Symbol:   "YLC",
 		Value:    "210572645",
@@ -62,30 +43,10 @@ var nativeTransferDst1 = blockatlas.Tx{
 	},
 }
 
-var nativeTransferDst2 = blockatlas.Tx{
-	ID:     "95CF63FAA27579A9B6AF84EF8B2DFEAC29627479E9C98E7F5AE4535E213FA4D0",
-	Coin:   coin.BNB,
-	From:   "tbnb1ttyn4csghfgyxreu7lmdu3lcplhqhxtzced45a",
-	To:     "tbnb12hlquylu78cjylk5zshxpdj6hf3t0tahwjt3ex",
-	Fee:    "125000",
-	Date:   1555117625,
-	Block:  7928667,
-	Status: blockatlas.StatusCompleted,
-	Memo:   "test",
-	Meta: blockatlas.NativeTokenTransfer{
-		TokenID:  "YLC-D8B",
-		Symbol:   "YLC",
-		Value:    "210572645",
-		Decimals: 8,
-		From:     "tbnb1ttyn4csghfgyxreu7lmdu3lcplhqhxtzced45a",
-		To:       "tbnb12hlquylu78cjylk5zshxpdj6hf3t0tahwjt3ex",
-	},
-}
-
-var utxoTransferDst1 = blockatlas.Tx{
+var utxoTransferDst1 = Tx{
 	ID:   "zpub6ruK9k6YGm8BRHWvTiQcrEPnFkuRDJhR7mPYzV2LDvjpLa5CuGgrhCYVZjMGcLcFqv9b2WvsFtY2Gb3xq8NVq8qhk9veozrA2W9QaWtihrC",
 	Coin: coin.BTC,
-	Inputs: []blockatlas.TxOutput{
+	Inputs: []TxOutput{
 		{
 			Address: "bc1qhn03cww757mnnlpkdvvfkaydxqygm86nvkm92h",
 			Value:   "1",
@@ -99,7 +60,7 @@ var utxoTransferDst1 = blockatlas.Tx{
 			Value:   "1",
 		},
 	},
-	Outputs: []blockatlas.TxOutput{
+	Outputs: []TxOutput{
 		{
 			Address: "bc1qjcslq88cht8llqmh3aqshjx9we9msv386jvxl6",
 			Value:   "3",
@@ -108,14 +69,14 @@ var utxoTransferDst1 = blockatlas.Tx{
 	Fee:    "125000",
 	Date:   1555117625,
 	Block:  592400,
-	Status: blockatlas.StatusCompleted,
+	Status: StatusCompleted,
 	Memo:   "test",
 }
 
-var utxoTransferDst2 = blockatlas.Tx{
+var utxoTransferDst2 = Tx{
 	ID:   "zpub6ruK9k6YGm8BRHWvTiQcrEPnFkuRDJhR7mPYzV2LDvjpLa5CuGgrhCYVZjMGcLcFqv9b2WvsFtY2Gb3xq8NVq8qhk9veozrA2W9QaWtihrC",
 	Coin: coin.BTC,
-	Inputs: []blockatlas.TxOutput{
+	Inputs: []TxOutput{
 		{
 			Address: "bc1q6e8sdxlgc7ekqkqyevtrx8wshfv7sg66z3z6ce",
 			Value:   "4",
@@ -125,7 +86,7 @@ var utxoTransferDst2 = blockatlas.Tx{
 			Value:   "2",
 		},
 	},
-	Outputs: []blockatlas.TxOutput{
+	Outputs: []TxOutput{
 		{
 			Address: "bc1q2fpry7zwqh575huc9urwfdvjtuvz508wez56ff",
 			Value:   "3",
@@ -142,32 +103,12 @@ var utxoTransferDst2 = blockatlas.Tx{
 	Fee:    "125000",
 	Date:   1555117625,
 	Block:  592400,
-	Status: blockatlas.StatusCompleted,
+	Status: StatusCompleted,
 	Memo:   "test",
 }
 
-var txsBlock = blockatlas.Block{
-	Number: 12345,
-	ID:     "12345",
-	Txs: []blockatlas.Tx{
-		transferDst1,
-		transferDst2,
-		nativeTransferDst1,
-		nativeTransferDst2,
-	},
-}
-
-func TestGetTxs(t *testing.T) {
-	txs := observer.GetTxs(&txsBlock)
-	assert.Equal(t, len(txs), 4)
-	assert.Equal(t, txs["tbnb1fhr04azuhcj0dulm7ka40y0cqjlafwae9k9gk2"].Size(), 2)
-	assert.Equal(t, txs["tbnb1sylyjw032eajr9cyllp26n04300qzzre38qyv5"].Size(), 1)
-	assert.Equal(t, txs["tbnb1ttyn4csghfgyxreu7lmdu3lcplhqhxtzced45a"].Size(), 2)
-	assert.Equal(t, txs["tbnb12hlquylu78cjylk5zshxpdj6hf3t0tahwjt3ex"].Size(), 2)
-}
-
 func TestTxSet_Add(t *testing.T) {
-	set := blockatlas.TxSet{}
+	set := TxSet{}
 	set.Add(&transferDst1)
 	var txs = set.Txs()
 	assert.Equal(t, txs[0].ID, transferDst1.ID)
